Authenticated Unit Routes
Must be authenticated to use all routes in unit. Must be the owner to create, edit, or delete a unit.

+++ backend/routers/unit.py
@@ -5,6 +5,11 @@ from backend.database import unit as unit_db
 from backend.dependencies import DBSession
 from backend.models import ClientErrorResponse, UnitResponse, UnitUpdate, UnitModule, CreateModule, ModuleDay
 
+from typing import Annotated
+from fastapi import Depends
+from backend.auth import get_firebase_user_from_token
+from backend.exceptions import UnauthorizedException
+
 router = APIRouter(prefix="/unit", tags=["Unit"])
 
 
@@ -14,8 +19,10 @@ router = APIRouter(prefix="/unit", tags=["Unit"])
             responses={
                  404: {"model": ClientErrorResponse}
              },
-            summary="Retrieve the modules and days for a given unit.")
-def get_unit_modules(unitID: int, session: DBSession) -> UnitResponse:
+            summary="Retrieve the modules and days for a given unit. Must be an authenticated user.")
+def get_unit_modules(unitID: int, 
+                     user: Annotated[dict, Depends(get_firebase_user_from_token)],
+                     session: DBSession) -> UnitResponse:
     """ Retrieve the modules and days for a given unit.
  
     Args:
@@ -29,7 +36,7 @@ def get_unit_modules(unitID: int, session: DBSession) -> UnitResponse:
         UnitResponse: A response model containing the modules and their days.
     """
     db_modules = unit_db.get_unit_modules(unitID, session) 
-
+    
     modules = []
     for m in db_modules:
         modules.append(UnitModule(
@@ -48,8 +55,10 @@ def get_unit_modules(unitID: int, session: DBSession) -> UnitResponse:
                  404: {"model": ClientErrorResponse},
                  409: {"model": ClientErrorResponse},
              },
-            summary="Create a new module for a given unit.")
-def create_new_module(unitID: int, module: CreateModule, session: DBSession) -> UnitResponse:
+            summary="Create a new module for a given unit. Must be an authenticated owner of the unit.")
+def create_new_module(unitID: int, 
+                      user: Annotated[dict, Depends(get_firebase_user_from_token)],
+                      module: CreateModule, session: DBSession) -> UnitResponse:
     """ 
     Create a new module for a given unit.
  
@@ -65,8 +74,13 @@ def create_new_module(unitID: int, module: CreateModule, session: DBSession) -> 
     Returns:
         UnitResponse: A response model containing the name and ID of the created module.
     """
+    user_id = user["uid"] 
+    teacherID = unit_db.get_teacher_id_by_unit_id(unitID, session)
+    if user_id != teacherID:
+        raise UnauthorizedException("create module")
+    
     module = unit_db.create_new_module(unitID, module, session)
-    return get_unit_modules(unitID, session)
+    return get_unit_modules(unitID, user, session)
 
 
 
@@ -74,8 +88,10 @@ def create_new_module(unitID: int, module: CreateModule, session: DBSession) -> 
             status_code= 204,
             responses={404: {"model": ClientErrorResponse},
                        422: {"model": ClientErrorResponse}},
-            summary="Update a unit's name and/or settings.")
-def update_unit(unitID: int, unitUpdate: UnitUpdate, session: DBSession):
+            summary="Update a unit's name and/or settings. Must be an authenticated owner of the unit.")
+def update_unit(unitID: int, 
+                user: Annotated[dict, Depends(get_firebase_user_from_token)],
+                unitUpdate: UnitUpdate, session: DBSession):
     """
     Update a unit's name and/or settings.
 
@@ -91,6 +107,11 @@ def update_unit(unitID: int, unitUpdate: UnitUpdate, session: DBSession):
     Returns:
         None
     """
+    user_id = user["uid"] 
+    teacherID = unit_db.get_teacher_id_by_unit_id(unitID, session)
+    if user_id != teacherID:
+        raise UnauthorizedException("edit module")
+    
     unit_db.update_unit(unitID, unitUpdate, session)
     
 
@@ -99,8 +120,10 @@ def update_unit(unitID: int, unitUpdate: UnitUpdate, session: DBSession):
 @router.delete("/{unitID}",
                status_code=204,
                responses={404: {"model": ClientErrorResponse}},
-               summary="Delete a unit.")
-def delete_classroom(unitID: int, session: DBSession):
+               summary="Delete a unit. Must be an authenticated owner of the unit.")
+def delete_classroom(unitID: int, 
+                     user: Annotated[dict, Depends(get_firebase_user_from_token)],
+                     session: DBSession):
     """Delete a unit by its ID.
     Args:
         unitID (int): The ID of the unit to delete.
@@ -110,4 +133,9 @@ def delete_classroom(unitID: int, session: DBSession):
     Returns:
         None
     """
+    user_id = user["uid"] 
+    teacherID = unit_db.get_teacher_id_by_unit_id(unitID, session)
+    if user_id != teacherID:
+        raise UnauthorizedException("delete module")
+    
     unit_db.delete_unit(unitID, session)
